generator: change is_isl_type to check for a registered isl_class
This changes is_isl_type to look for a class registered in the generator
instead of matching a 4 character name prefix on the type name.

As we only check for classes the method gets moved to a new method is_isl_class.
is_isl_type will be expanded to support checking for enums in the next
patch.

Signed-off-by: Andreas Simbuerger <simbuerg@fim.uni-passau.de>